backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
redis
rq
python-multipart
cryptography
python-jose[cryptography]
passlib[bcrypt]
pymysql
python-dotenv
